requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
mwparserfromhell==0.5.4
nltk==3.5
numpy==1.19.2
pkg-resources==0.0.0
python-Levenshtein==0.12.0
regex==2020.7.14
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wikitextparser==0.44.1
xgboost==1.2.0
